Changed Database

+++ server/votes/vote_repository.go
@@ -13,7 +13,7 @@ var ErrNoRowsAvailable = errors.New("No rows available")
 type VoteRepository interface {
 	VoteAPost(user_id, post_id int, vote_type string) error
 	GetPostVotes(postId int) ([]PostVotes, error)
-	DeletePostVotes(postId int) error
+	DeletePostVotes(postId, userId int) error
 	VoteAComment(user_id, comment_id int, vote_type string) error
 	GetPostComments(commentId int) ([]CommentVotes, error)
 	DeletePostComment(commentId int) error
@@ -31,11 +31,11 @@ func (r *SQLVoteRepository) VoteAPost(user_id, post_id int, vote_type string) er
 	if vote_type != "up" && vote_type != "down" {
         return ErrInvalidVoteType
     }
-	queryStatement := `INSERT INTO votes (user_id, post_id, vote_type) VALUES (?, ?, ?)`
+	queryStatement := `INSERT INTO votes (user_id, post_id, vote_type) VALUES ($1, $2, $3)`
 
 	_, err := r.db.Exec(queryStatement, user_id, post_id, vote_type)
 	if err != nil {
-		if strings.Contains(err.Error(), "UNIQUE constraint failed") {
+		if strings.Contains(err.Error(), "duplicate key value violates unique constraint") {
     	return ErrDuplicateVotes
 	}
 		return err
@@ -51,8 +51,8 @@ func (r *SQLVoteRepository) GetPostVotes(postId int) ([]PostVotes, error) {
 		COUNT(DISTINCT v.post_id) AS vote_count
 		FROM votes AS v
 		INNER JOIN users AS u ON v.user_id = u.id
-		WHERE v.post_id = ?
-		GROUP BY v.post_id
+		WHERE v.post_id = $1
+		GROUP BY v.post_id, u.username
 		ORDER BY v.created_at DESC
 	`
 
@@ -80,12 +80,12 @@ func (r *SQLVoteRepository) GetPostVotes(postId int) ([]PostVotes, error) {
 	return votes, nil
 }
 
-func (r *SQLVoteRepository) DeletePostVotes(postId int) error {
+func (r *SQLVoteRepository) DeletePostVotes(postId, userId int) error {
 	queryStatement := `
-		DELETE FROM votes WHERE post_id = ?
+		DELETE FROM votes WHERE post_id = $1 AND user_id = $2
 	`
 
-	result, err := r.db.Exec(queryStatement, postId)
+	result, err := r.db.Exec(queryStatement, postId, userId)
 	if err != nil {
 		return err
 	}
@@ -107,13 +107,13 @@ func (r *SQLVoteRepository) VoteAComment(user_id, comment_id int, vote_type stri
         return ErrInvalidVoteType
     }
 	queryStatement := `
-	INSERT INTO comment_votes (user_id, comment_id, vote_type) 
-	VALUES (?, ?, ?)
+		INSERT INTO comment_votes (user_id, comment_id, vote_type) 
+		VALUES ($1, $2, $3)
 	`
 
 	_, err := r.db.Exec(queryStatement, user_id, comment_id, vote_type)
 	if err != nil {
-		if strings.Contains(err.Error(), "UNIQUE constraint failed") {
+		if strings.Contains(err.Error(), "duplicate key value violates unique constraint") {
     	return ErrDuplicateVotes
 	}
 		return err
@@ -129,8 +129,8 @@ func (r *SQLVoteRepository) GetPostComments(commentId int) ([]CommentVotes, erro
 		COUNT(DISTINCT v.comment_id) AS vote_count
 		FROM comment_votes AS v
 		INNER JOIN users AS u ON v.user_id = u.id
-		WHERE v.comment_id = ?
-		GROUP BY v.comment_id
+		WHERE v.comment_id = $1
+		GROUP BY v.comment_id, u.username
 		ORDER BY v.created_at DESC
 	`
 
@@ -151,7 +151,7 @@ func (r *SQLVoteRepository) GetPostComments(commentId int) ([]CommentVotes, erro
 		commentVotes = append(commentVotes, vote)
 	}
 	
-	err = rows.Scan()
+	err = rows.Err()
 	if err != nil {
 		return nil, err
 	}
@@ -161,7 +161,7 @@ func (r *SQLVoteRepository) GetPostComments(commentId int) ([]CommentVotes, erro
 
 func (r *SQLVoteRepository) DeletePostComment(commentId int) error {
 	queryStatement := `
-		DELETE FROM comment_votes WHERE comment_id = ?
+		DELETE FROM comment_votes WHERE comment_id = $1
 	`
 
 	result, err := r.db.Exec(queryStatement, commentId)
